move logic to self contained generator
