+++ src/Maker/MakeAuthenticator.php
@@ -24,7 +24,7 @@ use Symfony\Bundle\MakerBundle\Security\SecurityConfigUpdater;
 use Symfony\Bundle\MakerBundle\Security\SecurityControllerBuilder;
 use Symfony\Bundle\MakerBundle\Str;
 use Symfony\Bundle\MakerBundle\Util\ClassSourceManipulator;
-use Symfony\Bundle\MakerBundle\Util\UseStatementCollection;
+use Symfony\Bundle\MakerBundle\Util\UseStatementGenerator;
 use Symfony\Bundle\MakerBundle\Util\YamlManipulationFailedException;
 use Symfony\Bundle\MakerBundle\Util\YamlSourceManipulator;
 use Symfony\Bundle\MakerBundle\Validator;
@@ -251,7 +251,7 @@ final class MakeAuthenticator extends AbstractMaker
 
     private function generateAuthenticatorClass(array $securityData, string $authenticatorType, string $authenticatorClass, $userClass, $userNameField): void
     {
-        $useStatements = new UseStatementCollection([
+        $useStatements = new UseStatementGenerator([
             Request::class,
             Response::class,
             TokenInterface::class,
@@ -268,7 +268,7 @@ final class MakeAuthenticator extends AbstractMaker
             $this->generator->generateClass(
                 $authenticatorClass,
                 'authenticator/EmptyAuthenticator.tpl.php',
-                ['use_statements' => $this->getFormattedUseStatements($useStatements)]
+                ['use_statements' => $useStatements->generateUseStatements()]
             );
 
             return;
@@ -294,7 +294,7 @@ final class MakeAuthenticator extends AbstractMaker
             $authenticatorClass,
             'authenticator/LoginFormAuthenticator.tpl.php',
             [
-                'use_statements' => $this->getFormattedUseStatements($useStatements),
+                'use_statements' => $useStatements->generateUseStatements(),
                 'user_fully_qualified_class_name' => trim($userClassNameDetails->getFullName(), '\\'),
                 'user_class_name' => $userClassNameDetails->getShortName(),
                 'username_field' => $userNameField,
@@ -315,7 +315,7 @@ final class MakeAuthenticator extends AbstractMaker
         );
 
         if (!class_exists($controllerClassNameDetails->getFullName())) {
-            $useStatements = new UseStatementCollection([
+            $useStatements = new UseStatementGenerator([
                 AbstractController::class,
                 Route::class,
                 AuthenticationUtils::class,
@@ -324,7 +324,7 @@ final class MakeAuthenticator extends AbstractMaker
             $controllerPath = $this->generator->generateController(
                 $controllerClassNameDetails->getFullName(),
                 'authenticator/EmptySecurityController.tpl.php',
-                ['use_statements' => $this->getFormattedUseStatements($useStatements)]
+                ['use_statements' => $useStatements->generateUseStatements()]
             );
 
             $controllerSourceCode = $this->generator->getFileContentsForPendingOperation($controllerPath);
