+++ src/Maker/MakeFunctionalTest.php
@@ -17,7 +17,7 @@ use Symfony\Bundle\MakerBundle\ConsoleStyle;
 use Symfony\Bundle\MakerBundle\DependencyBuilder;
 use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
-use Symfony\Bundle\MakerBundle\Util\UseStatementCollection;
+use Symfony\Bundle\MakerBundle\Util\UseStatementGenerator;
 use Symfony\Component\BrowserKit\History;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -64,7 +64,7 @@ class MakeFunctionalTest extends AbstractMaker
 
         $pantherAvailable = trait_exists(PantherTestCaseTrait::class);
 
-        $useStatements = new UseStatementCollection([
+        $useStatements = new UseStatementGenerator([
             ($pantherAvailable ? PantherTestCase::class : WebTestCase::class),
         ]);
 
@@ -72,7 +72,7 @@ class MakeFunctionalTest extends AbstractMaker
             $testClassNameDetails->getFullName(),
             'test/Functional.tpl.php',
             [
-                'use_statements' => $this->getFormattedUseStatements($useStatements),
+                'use_statements' => $useStatements->generateUseStatements(),
                 'web_assertions_are_available' => trait_exists(WebTestAssertionsTrait::class),
                 'panther_is_available' => $pantherAvailable,
             ]
