+++ src/Maker/MakeMessage.php
@@ -16,7 +16,7 @@ use Symfony\Bundle\MakerBundle\DependencyBuilder;
 use Symfony\Bundle\MakerBundle\FileManager;
 use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
-use Symfony\Bundle\MakerBundle\Util\UseStatementCollection;
+use Symfony\Bundle\MakerBundle\Util\UseStatementGenerator;
 use Symfony\Bundle\MakerBundle\Util\YamlSourceManipulator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -103,7 +103,7 @@ final class MakeMessage extends AbstractMaker
             'message/Message.tpl.php'
         );
 
-        $useStatements = new UseStatementCollection([
+        $useStatements = new UseStatementGenerator([
             MessageHandlerInterface::class,
             $messageClassNameDetails->getFullName(),
         ]);
@@ -112,7 +112,7 @@ final class MakeMessage extends AbstractMaker
             $handlerClassNameDetails->getFullName(),
             'message/MessageHandler.tpl.php',
             [
-                'use_statements' => $this->getFormattedUseStatements($useStatements),
+                'use_statements' => $useStatements->generateUseStatements(),
                 'message_class_name' => $messageClassNameDetails->getShortName(),
             ]
         );
