+++ src/Maker/MakeSerializerEncoder.php
@@ -15,7 +15,7 @@ use Symfony\Bundle\MakerBundle\ConsoleStyle;
 use Symfony\Bundle\MakerBundle\DependencyBuilder;
 use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
-use Symfony\Bundle\MakerBundle\Util\UseStatementCollection;
+use Symfony\Bundle\MakerBundle\Util\UseStatementGenerator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -56,7 +56,7 @@ final class MakeSerializerEncoder extends AbstractMaker
         );
         $format = $input->getArgument('format');
 
-        $useStatements = new UseStatementCollection([
+        $useStatements = new UseStatementGenerator([
             DecoderInterface::class,
             EncoderInterface::class,
         ]);
@@ -65,7 +65,7 @@ final class MakeSerializerEncoder extends AbstractMaker
             $encoderClassNameDetails->getFullName(),
             'serializer/Encoder.tpl.php',
             [
-                'use_statements' => $this->getFormattedUseStatements($useStatements),
+                'use_statements' => $useStatements->generateUseStatements(),
                 'format' => $format,
             ]
         );
