+++ src/Maker/MakeTwigExtension.php
@@ -15,7 +15,7 @@ use Symfony\Bundle\MakerBundle\ConsoleStyle;
 use Symfony\Bundle\MakerBundle\DependencyBuilder;
 use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
-use Symfony\Bundle\MakerBundle\Util\UseStatementCollection;
+use Symfony\Bundle\MakerBundle\Util\UseStatementGenerator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -55,7 +55,7 @@ final class MakeTwigExtension extends AbstractMaker
             'Extension'
         );
 
-        $useStatements = new UseStatementCollection([
+        $useStatements = new UseStatementGenerator([
             AbstractExtension::class,
             TwigFilter::class,
             TwigFunction::class,
@@ -64,7 +64,7 @@ final class MakeTwigExtension extends AbstractMaker
         $generator->generateClass(
             $extensionClassNameDetails->getFullName(),
             'twig/Extension.tpl.php',
-            ['use_statements' => $this->getFormattedUseStatements($useStatements)]
+            ['use_statements' => $useStatements->generateUseStatements()]
         );
 
         $generator->writeChanges();
